requirements.txt
Flask
boto3
python-dateutil
requests
